collision functions test
Here are some functions I was testing for collisions between balls with each other and with the walls, as well as an attempt at a function that simulates the complete system of two balls. It doesn't quite work yet because of the issue with repeated collisions, but we can work on that.